docs(compensation): add compesation events
The documentation cover the following points:

How to trigger compensation
Trigger compensation for the scope vs. for a given activity
Trigger compensation from an event subprocess
How to compensate an activity with a compensation handler
Under which conditions is a compensation handler invoked and how often
How compensation works for subprocesses
How compensation works with multi-instance
No compensation for child process instances
When is a compensation handler interrupted
Explain the compensation order (i.e. all compensation handlers at once) and how to enforce a specific order